backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
google-generativeai==0.3.1
requests==2.31.0
Pillow==10.0.0
sqlparse==0.4.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
